Clean up event handling
* No need to wait a second before adding the mousedown/keydown handlers.
It's unreliable due to the timeout and still results in our own handler
coming before non-capturing ones. Besides, we want our handler to take
precedence in most cases anyway due to poorly behaved websites. And the
main use-case (video elements) is no longer an issue now that we don't
do anything with the left and right keys (see previous commit).

* Make event listener cleanup automatic