Remove check for offscreen elements in visibleInDom
This was buggy because getBoundingClientRect is adjusted for scroll
position, so when scrolling the body of a document, the extension would
stop thinking the root node is visible.

This could be addressed by adjusting the value based on the scroll
offsets. But that would actually need to be done recursively, because
the value is effected by every ancestor with a scroll offset. Doing
that would be a pain and slow down the code, so I'll be finding a better
solution to the original problem.

Note that the code was also buggy because the height and bottom
properties didn't take overflow into account (though that also could
be fixed).